Galaxy refactoring to allow better reuse
* Change `Galaxy` class to take a specific set of parameters rather
  than `Option`s from `optparse`
* Extract `exit_without_ignore` from `GalaxyRole` class and move
  to `RoleRequirement` class
* Extract `read_roles_file` from `GalaxyRole` class and move to
  `RoleRequirement` class
* Create `install_roles` routine inside RoleRequirement module
